Drop the need to run "test" cross sbt versions
.. by reworking & moving ProblemReportingSpec from sbtplugin to core.
