+++ sbtplugin/src/main/scala/com/typesafe/tools/mima/plugin/SbtMima.scala
@@ -83,28 +83,8 @@ object SbtMima {
     if (failOnProblem && (backErrors.nonEmpty || forwErrors.nonEmpty)) sys.error(projectName + ": Binary compatibility check failed!")
   }
 
-  private[mima] def isReported(module: ModuleID, filters: Seq[core.ProblemFilter], versionedFilters: Map[String, Seq[core.ProblemFilter]])(log: Logging, projectName: String)(problem: core.Problem) = {
-
-    // version string "x.y.z" is converted to an Int tuple (x, y, z) for comparison
-    val versionOrdering = Ordering[(Int, Int, Int)].on { version: String =>
-      val ModuleVersion = """(\d+)\.?(\d+)?\.?(.*)?""".r
-      val ModuleVersion(epoch, major, minor) = version
-      val toNumeric = (revision: String) => Try(revision.replace("x", Short.MaxValue.toString).filter(_.isDigit).toInt).getOrElse(0)
-      (toNumeric(epoch), toNumeric(major), toNumeric(minor))
-    }
-
-    (versionedFilters.collect {
-      // get all filters that apply to given module version or any version after it
-      case f @ (version, versionFilters) if versionOrdering.gteq(version, module.revision) => versionFilters
-    }.flatten ++ filters).forall { f =>
-      if (f(problem)) {
-        true
-      } else {
-        log.debugLog(projectName + ": filtered out: " + problem.description + "\n  filtered by: " + f)
-        false
-      }
-    }
-  }
+  private[mima] def isReported(module: ModuleID, filters: Seq[core.ProblemFilter], versionedFilters: Map[String, Seq[core.ProblemFilter]])(log: Logging, projectName: String)(problem: core.Problem) =
+    ProblemReporting.isReported(module.revision, filters, versionedFilters)(log, projectName)(problem)
 
   /** Resolves an artifact representing the previous abstract binary interface
    *  for testing.
